requirements.txt
Flask
Celery
pymssql
SQLAlchemy
requests
pandas
pillow
redis
